mountinfo: Add stubs for non-Linux platforms
