+++ pkg/mountinfo/mountinfo_privileged_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// +build privileged_tests
+// +build linux,privileged_tests
 
 package mountinfo
